more lib packages + refactoring
