+++ src/components/Channel.vue
@@ -71,7 +71,7 @@
 
 <script lang="ts">
 import { Vue, Component, Prop } from 'vue-property-decorator';
-import { range, scale, clamp } from '@/utils';
+import { range, scale, clamp } from '@/lib/std';
 import { AnyEffect } from '@/core/filters/effect';
 import { Channel as C } from '@/core/channel';
 import { EffectMap, EffectName } from '@/core';
