+++ src/components/Editable.vue
@@ -11,7 +11,7 @@
 </template>
 
 <script lang="ts">
-import { Keys } from '@/utils';
+import { Keys } from '@/lib/std';
 import { update } from '@/lib/vutils';
 import { onMounted, ref, createComponent, watch } from '@vue/composition-api';
 
@@ -71,7 +71,7 @@ export default createComponent({
     }
 
     function keydown(e: KeyboardEvent) {
-      if (e.keyCode !== Keys.ENTER) { return; }
+      if (e.keyCode !== Keys.Return) { return; }
       if (el.value) {
         el.value.blur();
       }
