+++ src/core/scheduled/automation.ts
@@ -3,7 +3,7 @@ import { AutomationClip } from '@/core/automation';
 import * as Audio from '@/lib/audio';
 import { Schedulable, SchedulableType } from '@/core/scheduled/schedulable';
 import { Serializable } from '@/core/serializable';
-import { literal } from '@/utils';
+import { literal } from '@/lib/std';
 
 export const ScheduledAutomationType = t.intersection([
   t.type({
