+++ src/dawg/extensions/core/notify/Notifications.vue
@@ -42,7 +42,7 @@
 
 <script lang="ts">
 import { Vue, Component, Prop} from 'vue-property-decorator';
-import { reverse } from '@/utils';
+import { reverse } from '@/lib/std';
 import * as framework from '@/lib/framework';
 import { Marked } from 'marked-ts';
