+++ src/dawg/extensions/extra/mixer/Mixer.vue
@@ -23,7 +23,7 @@
 import { Component, Prop, Vue } from 'vue-property-decorator';
 import Channel from '@/components/Channel.vue';
 import { Watch } from '@/update';
-import { range } from '@/utils';
+import { range } from '@/lib/std';
 import { AnyEffect, Channel as C } from '@/core';
 
 @Component({
