+++ src/lib/framework/SideTabs.vue
@@ -38,8 +38,8 @@
 </template>
 
 <script lang="ts">
-import BaseTabs from '@/components/BaseTabs.vue';
-import Tab from '@/components/Tab.vue';
+import BaseTabs from '@/lib/framework/BaseTabs.vue';
+import Tab from '@/lib/framework/Tab.vue';
 import * as framework from '@/lib/framework';
 import { createComponent, computed } from '@vue/composition-api';
