+++ src/lib/framework/index.ts
@@ -4,6 +4,7 @@ export * from '@/lib/framework/extensions';
 export * from '@/lib/framework/menu';
 export * from '@/lib/framework/manager';
 export * from '@/lib/framework/constants';
+export * from '@/lib/framework/platform';
 
 
 import * as notify from '@/lib/framework/notify';
