+++ src/sequencer/AutomationClipElement.vue
@@ -35,7 +35,7 @@
 <script lang="ts">
 import { Vue, Component, Prop, Inject } from 'vue-property-decorator';
 import { Point, ScheduledAutomation } from '@/core';
-import { scale } from '@/utils';
+import { scale } from '@/lib/std';
 import * as dawg from '@/dawg';
 
 @Component
