+++ src/utils.ts
@@ -1,19 +1,6 @@
-type Color = 'black' | 'white';
+import { range } from '@/lib/std';
 
-export const range = (a: number, b = 0, interval = 1) => {
-  let start;
-  let end;
-  if (a > b) {
-    start = 0; end = a;
-  } else {
-    start = a; end = b;
-  }
-  const rge = [];
-  for (let i = start; i < end; i += interval) {
-    rge.push(i);
-  }
-  return rge;
-};
+type Color = 'black' | 'white';
 
 interface OctaveKey {
   value: string;
@@ -40,32 +27,6 @@ range(0, 9).reverse().forEach((value) => {
   });
 });
 
-export const copy = <T>(o: T): T => {
-  return JSON.parse(JSON.stringify(o));
-};
-
-export function mapRange(x: number, inMin: number, inMax: number, outMin: number, outMax: number) {
-  return (((x - inMin) * (outMax - outMin)) / (inMax - inMin)) + outMin;
-}
-
-export const Keys = {
-  ENTER: 13,
-  SHIFT: 16,
-  DELETE: 46,
-  BACKSPACE: 8,
-  SPACE: 32,
-  UP: 38,
-  DOWN: 40,
-  RIGHT: 39,
-  LEFT: 37,
-};
-
-export const Button = {
-  LEFT: 0,
-  MIDDLE: 1,
-  RIGHT: 2,
-};
-
 export const findUniqueName = (objects: Array<{ name: string }>, prefix: string) => {
   let name: string;
   let count = 1;
@@ -89,17 +50,6 @@ export const findUniqueName = (objects: Array<{ name: string }>, prefix: string)
   return name;
 };
 
-export function scale(value: number, from: [number, number], to: [number, number]) {
-  return (value - from[0]) * (to[1] - to[0]) / (from[1] - from[0]) + to[0];
-}
-
-export function clamp(value: number, min: number, max: number) {
-  return Math.max(min, Math.min(max, value));
-}
-
-
-export type ConstructorOf<T> = new (...args: any[]) => T;
-
 export const disposeHelp = (o: { dispose: () => void }) => {
   // Tone.js
   try {
@@ -108,129 +58,4 @@ export const disposeHelp = (o: { dispose: () => void }) => {
     // tslint:disable-next-line:no-console
     console.info(`dispose failed for ${o} =>`, e.message);
   }
-};
-
-export function* reverse <T>(arr: T[]) {
-  for (let i = arr.length - 1; i >= 0; i--) {
-    yield arr[i];
-  }
-}
-
-export type Primitive = string | number | boolean | undefined | null;
-
-export function literal<T extends Primitive>(value: T): T {
-  return value;
-}
-
-type Events = keyof WindowEventMap;
-
-type EventListener<K extends Events> = (ev: WindowEventMap[K]) => any;
-
-type EventListeners = {
-  [P in keyof WindowEventMap]?: EventListener<P> | 'remove';
-};
-
-/**
- * Add 0 or more event listeners and return an object with a dispose method to remove the listeners.
- *
- * @param events The events.
- * @param options The options.
- */
-export const addEventListeners = (
-  events: EventListeners,
-  options?: boolean | AddEventListenerOptions,
-) => {
-  const types = Object.keys(events) as Events[];
-
-  const remove = () => {
-    for (const type of types) {
-      const ev = events[type];
-      if (ev === 'remove') {
-        continue;
-      }
-
-      window.removeEventListener(type, ev as any);
-    }
-  };
-
-  for (const type of types) {
-    const ev = events[type];
-    if (ev === 'remove') {
-      // @ts-ignore
-      // There is a weird error with union types
-      // Going to just ignore this
-      events[type] = remove;
-    }
-    window.addEventListener(type, ev as any, options);
-  }
-
-
-  return {
-    dispose: remove,
-  };
-};
-
-/**
- * Add an event listener (like normal) but return an object with a dispose method to remove the same listener.
- *
- * @param type The event.
- * @param ev The listener.
- * @param options The options.
- */
-export const addEventListener = <K extends Events>(
-  type: K,
-  ev: EventListener<K>,
-  options?: boolean | AddEventListenerOptions,
-) => {
-  window.addEventListener(type, ev, options);
-
-  return {
-    dispose: () => {
-      window.removeEventListener(type, ev);
-    },
-  };
-};
-
-export const keys = <O>(o: O): Array<keyof O & string> => {
-  return Object.keys(o) as Array<keyof O & string>;
-};
-
-export const mapObject = <V, T, O extends { [k: string]: V }>(o: O, f: (v: V) => T) => {
-  const transformed: { [k: string]: T } = {};
-  for (const key of keys(o)) {
-    transformed[key] = f(o[key]);
-  }
-  return transformed as { [K in keyof O]: T };
-};
-
-export const uniqueId = () => {
-  return Math.random().toString().substr(2, 9);
-};
-
-export function* chain<T>(...arrays: T[][]) {
-  for (const array of arrays) {
-    for (const item of array) {
-      yield item;
-    }
-  }
-}
-
-export const makeLookup = <T extends { id: string }>(array: Iterable<T>) => {
-  const lookup: { [k: string]: T } = {};
-  for (const item of array) {
-    lookup[item.id] = item;
-  }
-  return lookup;
-};
-
-interface Ordered {
-  order?: number;
-}
-
-export const sortOrdered = (a: Ordered, b: Ordered) => {
-  return (a.order || 0) - (b.order || 0);
-};
-
-export const isDefined = <T>(t: T | undefined): t is T => {
-  return t !== undefined;
 };
